+++ probe-rs/src/architecture/arm/core/armv8a.rs
@@ -981,6 +981,12 @@ impl<'probe> CoreInterface for Armv8a<'probe> {
         }
         Ok(breakpoints)
     }
+
+    fn fpu_support(&mut self) -> Result<bool, crate::error::Error> {
+        Err(crate::error::Error::Other(anyhow::anyhow!(
+            "Fpu detection not yet implemented"
+        )))
+    }
 }
 
 impl<'probe> MemoryInterface for Armv8a<'probe> {