+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Support added for trace funnels and SWO peripherals
   - Added custom sequencing for STM32H7 parts to configure debug system components on attach
 - Added support for ARMv8-A cores running in 64-bit mode (#1120)
+- Added FPU register reading support for cortex-m cores
 
 ### Changed
 
@@ -62,6 +63,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated all parameters and fields that refer to memory addresses from u32 to u64 in preparation for 64-bit target support. (#1115)
 - Updated `Core::read_core_reg` and `Core::write_core_reg` to work with both 32 and 64-bit values (#1119)
 - Renamed `core::CoreRegisterAddress` to `core::RegisterId`, and `core::CoreRegister` to `core::MemoryMappedRegister`. (#1121)
+- Updated gdb-server to use gdbstub internally (#1125)
+- gdb-server now uses all cores on a target (#1125)
 
 ### Fixed
 
@@ -89,6 +92,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - probe_rs::debug::Registers uses new `core::RegisterId` and `core::RegisterValue` for consistent register handling.
   - RISCV `Disassembly` works correctly for 'compressed' (RV32C isa variants) instruction sets.
   - RISCV stack unwind improvements (stack frames and registers work, variables do not resolve correctly.)
+- Fixed a possible endless recursion in the J-Link code, when no chip is connected. (#1123)
 
 ## [0.12.0]